chore: Update environment variables and next.config
This commit updates the environment variables in the .env file to set the NEXT_PUBLIC_API_URL and NEXT_PUBLIC_IMAGE_URL to the local development URLs. Additionally, the next.config.mjs file is modified to include the remotePatterns configuration for the images module, allowing images from the localhost to be loaded.

These changes ensure that the application is configured correctly for local development and that the images module can handle remote patterns.